% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpolboot.R
\name{rdpolboot}
\alias{rdpolboot}
\title{rdpolboot}
\usage{
rdpolboot(y, x, c = 0, fuzzy = NULL, covs = NULL, kernel = "tri", deriv = NULL, 
cluster = NULL, scaleregul = 1, groups = NULL, groupreps = 5, 
mainpol = NULL, minpol = 1, maxpol = 4, reps = 15000, alpha = 0.05)
}
\arguments{
\item{y}{a vector of the dependent variable.}

\item{x}{a vector of the running variable (a.k.a. score or forcing variable).}

\item{c}{specifies the RD cutoff in x; default is \code{c = 0}.}

\item{fuzzy}{specifies the treatment status variable used to implement fuzzy RD estimation (or Fuzzy Kink RD if \code{deriv = 1} is also specified). Default is Sharp RD design and hence this option is not used.}

\item{covs}{specifies additional covariates to be used for estimation and inference.}

\item{kernel}{is the kernel function used to construct the local-polynomial estimator(s). Options are triangular (default option), epanechnikov and uniform.}

\item{deriv}{specifies the order of the derivative of the regression functions to be estimated. Default is \code{deriv = 0} (for Sharp RD, or for Fuzzy RD if fuzzy is also specified). Setting \code{deriv = 1} results in estimation of a Kink RD design (up to scale), or Fuzzy Kink RD if fuzzy is also specified.}

\item{cluster}{indicates the cluster ID variable used for cluster-robust variance estimation with degrees-of-freedom weights.}

\item{scaleregul}{specifies scaling factor for the regularization term added to the denominator of the bandwidth selectors. Setting \code{scaleregul = 0} removes the regularization term from the bandwidth selectors.  Default is \code{scaleregul = 1}.}

\item{alpha}{specifies the percentage level of the BCa confidence interval; default is \code{alpha = 0.05}, yielding a 95\% confidence interval.}

\item{mainpol}{specifies the main polynomial order to consider for robustness testing. If not specified, the polynomial order with the lowest AMSE is used as the main polynomial order; default is \code{mainpol = NULL}, and hence this option is not used.}

\item{groups}{specifies the number groups the data is to be randomly split in. This option speeds up the jackknife calculations for estimating the acceleration value in case the data set is very large. Note that it does not have to exactly divide the total number of observations. Be careful not to choose a very large number of groups, as too small group sizes can crash calculation of the AMSE. Default is \code{groups = NULL}, and hence this option is not used.}

\item{groupreps}{specifies the number of repetitions for the jackknife procedure; default is \code{groupreps = 1}. If \code{groups} is specified, it is recommended to increase \code{groupreps} for a more reliable approximation of the acceleration factor.}

\item{minpol}{specifies the lowest polynomial order to consider for estimating bootstrapped confidence intervals; default is \code{minpol = 1}.}

\item{maxpol}{specifies the highest polynomial order to consider for estimating bootstrapped confidence intervals; default is \code{maxpol = 4}.}

\item{reps}{specifies the number of bootstrap repetitions; default is \code{reps = 15000}.}
}
\value{
\item{N}{total number of observations}
\item{minpol}{lowest polynomial order considered for estimating bootstrapped confidence intervals.}
\item{maxpol}{highest polynomial order considered for estimating bootstrapped confidence intervals.}
\item{mainpol}{main polynomial order (the one with the lowest AMSE if not specified).}
\item{missings_bs}{number of unsuccessful bootstrap iterations.}
\item{missings_jk}{number of unsuccessful jackknife iterations.}
\item{amses}{AMSE value corresponding to each polynomial order.}
\item{bca_ci}{BCa confidence intervals for each polynomial order pair.}
\item{included}{list of the main polynomial order and all polynomial orders whose fit is indistinguishable.}
\item{plot}{plot of the BCa confidence intervals.}
}
\description{
rdpolboot estimates bootstrapped confidence intervals around the asymptotic mean squared error (AMSE) of different polynomial orders in a regression discontinuity design. These compare the fit of different polynomial orders and can be used to choose between them if the estimate is sensitive to the polynomial order choice.
The package builds upon rdbwselect, the CCT bandwidth selector from Calonico, Cattaneo and Titiunik (2014a, 2014b), to calculate the MSE-optimal bandwidth for each polynomial order. Second, it uses rdmse from Pei, Lee, Card and Weber (2021) to estimate the AMSEs for each bootstrapped sample.
}
\examples{
# Let y be the outcome variable, x the running variable, and t the treatment instrument:

# Estimation for sharp RD designs with default settings
out <- rdpolboot(y, x)

# Estimation for fuzzy RD designs with default settings
out <- rdpolboot(y, x, fuzzy = t)

# Estimation for sharp RD designs using groups and groupreps to speed up estimation
# of the acceleration factor for large data sets
out <- rdpolboot(y, x, groups = 5, groupreps = 40)

# Estimation for sharp RD designs with different min and max polynomials
out <- rdpolboot(y, x, minpol = 0, maxpol = 3)

# Estimation for sharp RD designs with covariates and a cluster variable
out <- rdpolboot(y, x, covs = covars, cluster = clustervar)

}
